# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Sneak  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/user": {
            "post": {
                "tags": [
                    "Users"
                ],
                "summary": "Crear nuevo user",
                "description": "Returns user data",
                "operationId": "createdUser",
                "requestBody": {
                    "description": "Datos para crear usuario",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "surname",
                                    "email",
                                    "address",
                                    "password"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Victor"
                                    },
                                    "surname": {
                                        "type": "string",
                                        "example": "Monzo Mora"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "victor.monzo.mora2@gmail.com"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "C/ Correos"
                                    },
                                    "password": {
                                        "type": "string",
                                        "example": "12345678"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/api/user/3": {
            "put": {
                "tags": [
                    "Users"
                ],
                "summary": "Editar user",
                "description": "Returns user data",
                "operationId": "editUser",
                "requestBody": {
                    "description": "Ponga sus credenciales",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "surname",
                                    "address"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Victor"
                                    },
                                    "surname": {
                                        "type": "string",
                                        "example": "Monzo Mora"
                                    },
                                    "address": {
                                        "type": "string",
                                        "example": "C/ Correos"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/User"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string",
                                            "example": "Usuario no autenticado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                },
                "security": [
                    {
                        "apiAuth": []
                    }
                ]
            }
        }
    },
    "components": {
        "schemas": {
...
            "User": {
                "required": [
                    "token, id, name, surname, email, address, created_at, photo"
                ],
                "properties": {
                    "token": {
                        "type": "string",
                        "readOnly": "true",
                        "example": "12 |afafafdeavafafr"
                    },
                    "name": {
                        "type": "string",
                        "readOnly": "true",
                        "example": "Victor"
                    },
                    "surname": {
                        "type": "string",
                        "readOnly": "true",
                        "example": "Monzo Mora"
                    },
                    "email": {
                        "type": "string",
                        "readOnly": "true",
                        "example": "victor.monzo.mora@gmail.com"
                    },
                    "address": {
                        "type": "string",
                        "readOnly": "true",
                        "example": "C/ Correos"
                    },
                    "created_at": {
                        "type": "string",
                        "readOnly": "true",
                        "example": "11/11/11"
                    },
                    "photo": {
                        "type": "string",
                        "readOnly": "true",
                        "example": "/imgs/products-users/DB/U-victor.jpg"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "User"
                }
            }
        },
...
        "securitySchemes": {
            "apiAuth": {
                "type": "http",
                "description": "Login with email and password to get the authentication token",
                "name": "Token based Based",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
    "security": [
        []
    ]
}